透明度混合中间交界处bug修改，鼠标键盘操作模式修改

+++ include/Utils.h
@@ -13,6 +13,6 @@ namespace Utils
 
 	extern glm::vec3 Barycentric(glm::vec2 A, glm::vec2 B, glm::vec2 C, glm::vec2 P);
 
-	extern glm::vec4 lerp(const glm::vec4 &A, const glm::vec4 &B, double t);
+	extern glm::vec4 lerp(const glm::vec4 &A, const glm::vec4 &B, float t);
 }
