+++ src/Casli/Core/DeviceContext.h
@@ -54,7 +54,7 @@ private:
 	
 	void ViewportTransform(std::vector<glm::vec4> vertex[3]);
 	unsigned char *Vertex(int idx, unsigned char *vertexBuffer);
-	void DDXDDY(std::vector<glm::vec4> vertex[3], glm::ivec3 &t0, glm::ivec3 &t1, glm::ivec3 &t2, glm::ivec2 &P);
+	void DDXDDY(std::vector<glm::vec4> vertex[3], glm::vec3 &t0, glm::vec3 &t1, glm::vec3 &t2, glm::ivec2 &P);
 	void prePerspCorrection(std::vector<glm::vec4> output[3]);
 	unsigned char * Interpolation(std::vector<glm::vec4> vertex[3], glm::vec3 &bcScreen);
 	//Alpha Blend
